exercicio_5

+++ Estrutura_de_Decisão/exercicio_005.kt
@@ -8,26 +8,26 @@ A mensagem "Aprovado com Distinção", se a média for igual a dez. */
 
 fun main(){
 
-    val nota1: Float
-    val nota2: Float
-    val media: Float
+    val nota1: Double
+    val nota2: Double
+    val media: Double
 
     println("Informe suas duas notas: ")
-    nota1 = readLine()!!.toFloat()
-    nota2 = readLine()!!.toFloat()
+    nota1 = readLine()!!.toDouble()
+    nota2 = readLine()!!.toDouble()
 
     media = (nota1 + nota2)/2
 
-    if(media >= 0 && media < 7){
+    if(media >= 0 && media < 7.0){
 
         println("REPROVADO !!")
     }
-    else if(media >= 7 && media < 10){
+    else if(media >= 7.0 && media < 10.0){
 
         println("APROVADO !!")
 
     }
-    else if(media == 10){
+    else if(media == 10.0){
 
         println("APROVADO COM DISTINÇÃO  !!")
     }
